Some refactoring; Alignment struct; simple CLI
I did some random refactoring and renaming, but nothing major.

I implemented a simple Alignment struct to help organize search result
reporting. It's probably going to change a lot.

I finally implemented a simple CLI that allows supplying the target and
query from the command line. It uses clap, which I like a lot, but it
turns out to add perhaps not insignificant bloat to the size of the
binary. I'll plan to investigate this later on.