reorganized main align loop, fixed RowBounds.reuse(), renamed CloudBoundGroup to AntiDiagonalBounds
I reorganized the main alignment loop such that it is much more clear
which bits of data are cloned (or Mutex'd and Arc'd) across threads.

The RowBounds.reuse() method was psychotic and not really suitable for
being used, so I fixed that.

I renamed CloudBoundGroup to AntiDiagonalGroup, since I think it's a
more descriptive name and fits better with its counterpart (RowBounds).