fixed a RowBounds bug, added bias to output, fixed prefilter bug
There was a bug in the RowBounds implementation that was causing a debug
assert to fail when a RowBounds struct was initialized.

The composition bias field was added to tabular output.

MMseqs2 has some very opinionated sequence renaming features, which was
causing huge issues when trying to map seeds to targets & profiles for
nail alignment. I (hopefully) fixed this for most cases by forcing
mmseqs align to place the entire sequence/profile headers in the
tabular output (instead of whatever mmeqs happened to decide the "name"
was).