reworked null two, fixed some bugs
The rework to the null two is an adaptation of HMMER's null two by
expectation that should work better for the sparse DP clouds. The idea
is to weight the contribution of each target position by the posterior
probability that the position was emitted by a core model state.

I also fixed a display bug in the mid line of the alignment output,
switched back to explicitly running mmseqs prefilter & align (instead of
just running mmseqs search), added a debug seed step for setting
explicit row bounds, and dropped some of the optimizations off the
release build to dramatically reduce compile times (for now).