restored CLI functionality, AD bounds bugfixes
This restructures the CLI to restore the nail seed command and to accept
seeds in nail search.

This also fixes some bugs related to AD bounds not being properly reset
in weird situations where the min/max AD indices became invalid. The
culprit was the trim_wings() function, which destroys the cloud in some
pathological cases. For now, the fix is to run join_merge() on fewer
clouds, depending on the results of the intersection() function that is
added in this commit.

This also adds some new AD bound functionality: advance_forward() and
advance_backward(), which will simplify some of the logic in cloud
search once I get around to refactoring it.