removed Hmm struct; search pipeline now uses P7Hmm
The Hmm struct was really just a remnant of porting from HMMER, and it
was never really used except to create a Profile struct, so I've
collapsed all of that.

The search pipeline now uses the new P7Hmm struct to stream profiles.
I still need to make sure everything is still working as expected.